+++ x/airdrop/keeper/keeper_test.go
@@ -1,88 +1,25 @@
 package keeper_test
 
 import (
-	"fmt"
-
-	"github.com/cosmos/cosmos-sdk/baseapp"
-	"github.com/cosmos/cosmos-sdk/simapp"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	minttypes "github.com/cosmos/cosmos-sdk/x/mint/types"
-	"github.com/cosmos/cosmos-sdk/x/staking"
-	"github.com/cosmos/cosmos-sdk/x/staking/teststaking"
-	"github.com/stretchr/testify/suite"
-	"github.com/tendermint/tendermint/crypto/secp256k1"
-	tmrand "github.com/tendermint/tendermint/libs/rand"
-	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 
-	ojoapp "github.com/ojo-network/ojo/app"
-	appparams "github.com/ojo-network/ojo/app/params"
-	"github.com/ojo-network/ojo/x/airdrop/keeper"
 	"github.com/ojo-network/ojo/x/airdrop/types"
 )
 
-const (
-	displayDenom string = appparams.DisplayDenom
-	bondDenom    string = appparams.BondDenom
+func (s *IntegrationTestSuite) TestGetAirdropAccount() {
+	app, ctx := s.app, s.ctx
 
-	initialPower = int64(10000000000)
-)
+	originAmount := sdk.MustNewDecFromStr("500")
+	airdropAccount := types.AirdropAccount{
+		OriginAddress: "test",
+		OriginAmount:  &originAmount,
+	}
+	err := app.AirdropKeeper.SetAirdropAccount(ctx, airdropAccount)
+	s.Require().NoError(err)
 
-// Test addresses
-var (
-	valPubKeys = simapp.CreateTestPubKeys(2)
+	airdropAccount2, err := app.AirdropKeeper.GetAirdropAccount(ctx, "test")
+	s.Require().NoError(err)
 
-	valPubKey = valPubKeys[0]
-	pubKey    = secp256k1.GenPrivKey().PubKey()
-	addr      = sdk.AccAddress(pubKey.Address())
-	valAddr   = sdk.ValAddress(pubKey.Address())
-
-	valPubKey2 = valPubKeys[1]
-	pubKey2    = secp256k1.GenPrivKey().PubKey()
-	addr2      = sdk.AccAddress(pubKey2.Address())
-	valAddr2   = sdk.ValAddress(pubKey2.Address())
-
-	initTokens = sdk.TokensFromConsensusPower(initialPower, sdk.DefaultPowerReduction)
-	initCoins  = sdk.NewCoins(sdk.NewCoin(appparams.BondDenom, initTokens))
-)
-
-type IntegrationTestSuite struct {
-	suite.Suite
-
-	ctx         sdk.Context
-	app         *ojoapp.App
-	queryClient types.QueryClient
-	msgServer   types.MsgServer
-}
-
-func (s *IntegrationTestSuite) SetupTest() {
-	require := s.Require()
-	isCheckTx := false
-	app := ojoapp.Setup(s.T())
-	ctx := app.BaseApp.NewContext(isCheckTx, tmproto.Header{
-		ChainID: fmt.Sprintf("test-chain-%s", tmrand.Str(4)),
-		Height:  9,
-	})
-
-	queryHelper := baseapp.NewQueryServerTestHelper(ctx, app.InterfaceRegistry())
-	types.RegisterQueryServer(queryHelper, keeper.NewQuerier(app.AirdropKeeper))
-
-	sh := teststaking.NewHelper(s.T(), ctx, *app.StakingKeeper)
-	sh.Denom = bondDenom
-	amt := sdk.TokensFromConsensusPower(100, sdk.DefaultPowerReduction)
-
-	// mint and send coins to validators
-	require.NoError(app.BankKeeper.MintCoins(ctx, minttypes.ModuleName, initCoins))
-	require.NoError(app.BankKeeper.SendCoinsFromModuleToAccount(ctx, minttypes.ModuleName, addr, initCoins))
-	require.NoError(app.BankKeeper.MintCoins(ctx, minttypes.ModuleName, initCoins))
-	require.NoError(app.BankKeeper.SendCoinsFromModuleToAccount(ctx, minttypes.ModuleName, addr2, initCoins))
-
-	sh.CreateValidator(valAddr, valPubKey, amt, true)
-	sh.CreateValidator(valAddr2, valPubKey2, amt, true)
-
-	staking.EndBlocker(ctx, *app.StakingKeeper)
-
-	s.app = app
-	s.ctx = ctx
-	s.queryClient = types.NewQueryClient(queryHelper)
-	s.msgServer = keeper.NewMsgServerImpl(app.AirdropKeeper)
+	s.Require().Equal(airdropAccount2.OriginAddress, airdropAccount.OriginAddress)
+	s.Require().Equal(airdropAccount2.OriginAmount, airdropAccount.OriginAmount)
 }
